Add wrapper to start rebar3 bsp with mapped fds
Since rebar3 is very hard to actually silence on fd 1, we now have an
option to ask rebar3 bsp to communicate on a given set of FDs. These
FDs are specified via the environment parameter REBAR3_BSP_IO_FDS,
which is scanned for three integers.

The given integers are used by rebar3_bsp_portio, which implements the
JSON-RPC protocol interaction.

The wrapper shell script rebinds fds as follows:

stdin  (0) -> 3
stdout (1) -> 4
stderr (2) -> 5
stdin  (0) <- /dev/null
stdout (1) <- /dev/null
stderr (2) <- /dev/null

Before invoking rebar3 bsp REBAR3_BSP_IO_FDS is set to "3 4 5". With
FDs 0,1,2 bound to /dev/null any spurious output produced by rebar3
now disappears into the void.

Note that rebar3 bsp currently does not use the third ("stderr") fd,
but the BSP protocol mentions that it is to be used to "to report
execution progress to the user". This will be challenging to support
in erlang_ls, as the handling of stderr is deficient in BEAM itself.